requirements.txt
django >= 3
psycopg2
requests
djangorestframework
django-cors-headers
django-environ